Refactor BuildStream kubernetes configuration
Updated BuildStream client configuration tafter replacing our project
built BuildStream Docker image witht the one produced by the BuildStream
team.

Previously the configuration was expecting the BuildStream repo to be
available in the docker image.
However since moving to the official BuildStream docker image the repo
is not part of the image.

The new approch uses ConfigMap to define all the files needed directly
in the kubernetes definition.