Refactor kubernetes BuildStream client config
Replaced the extensive use of configMaps with initContainer

The previous solution was creating the whole project using configMaps,
leading the yaml file to be quite long and forcing us to maintain the
whole project folder.

Now instead the CI pipeline will pull BuildStream repo in an empty
volume using an initContainer. Once this is done the main container
mount only the autotools folder (`doc/examples/autotools`)
Finally we need to append the buildgrid servers to the project.conf
file. This part still uses configMap