Add Jaerger monitoring to buildbarn cluster
Add Jaeger elements to the Buildbarn server definition. The following
elements are added:

+ ConfigMap
  + jaeger-configuration

+ Service
  + cassandra
  + jaeger-collector
  + jaeger-query

+ Deployment
  + jaeger-collector
  + jaeger-query

+ StatefulSet
  + cassandra

+ Job
  + jaeger-cassandra-schema-job

jaeger-collector gets traces from the different agent available and adds
it to the storage.
jaeger-query provides a web ui to access the traces recorded in the
storage.
cassandra is the DB used to store all traces.

This will create a Jaeger ecosystem in buildbarn namesapce recording
traces during execution. This will be taken down at the same time as the
rest of the buildbarn namesapce.
A decision must be made regarding what to do with the traces collected
and if we want to keep them.