Update version of bazel built during speed tests
This change is following the version update of Bazel used in the client.
The project moved from using Bazel 0.29.1 to 1.1.0.

As it moved to a new Major version of Bazel it cannot build versions of
Bazel prior to this major (o.x.x).
Prior to this change the bazel-build-speed-test was attempting to build
Bazel version 0.26.1. Therefore changing this version 1.1.0.