Add buildfarm/bazel testing with docker-compose
This adds the requisite docker-compose files for deploying
buildfarm, and any necessary configuration.

In addition, to make the bazel client interchangeable
between server implementations, the buildbarn configuration
is tweaked somewhat to eliminate the dependency on the
rbe 16.04 container. The run.sh script is also made
more configurable so that it can be used for more than
just buildbarn and bazel.

Gitlab-ci.yml jobs are added for buildfarm and bazel.