Build docker images from single Dockefile using buildkit
Currently, the images that are being created from separate
Dockerfiles, even though they often could be composed from
common layers, such as for defining a base build environment.
This is not ideal, as a build dependency that needs to be added for,
say the bazel client such as libpng, will need to be added
for all the worker images as well.

Therefore, define the construction of the images as a
multistage dockerfile. This has the downside, of now needing
to express all the image construction as a single file, but should
ultimately reduce the amount of boilerplate required. To reduce the
amount of unnecessary rebuilding of stages required, use buildkit
as the docker build backend instead.