Add goma client checking with docker-compose
This adds the goma client and server combination,
to test against the existing REAPI server
implementations.

To be compatible for testing and to facilitate
usage, some patches to the goma reapi server
backend have to be added. These are mainly focused
around authentication, input root handling and
platform message parsing.