Fix badge 404 for remote-asset test cases
This issue was caused by the usage of hyphons
within the CLIENT and ASSET variables, these are used
in the badge url which shields.io uses to delimit arguments.

Instead underscores should be used to split multiple words in
CLIENT, SERVER, ASSET variables. This has been updated in CONTRIBUTING.md.