dockerfile: put a python symlink in build_env
add a symlink from /usr/bin/python to /usr/bin/python3, in the build_env
docker image.
This symlink was already used in the Goma image, and will be a
requirement for building BuildFarm when we update to the newest
BuildFarm image.
By putting the symlink in build_env (the base image) we make it
available in all the docker images.