docker-compose-pants-yml: paramaterize PANTS_COMMIT
move some of the pants set-up commands into the Dockerfile, and make
PANTS_COMMIT one of the build args for the docker image.

leaves only one command defined in the docker-compose yaml file, which
therefore does not need to be surrounded by "sh -c".

also edit the auto-updater to update the new pants build arg